A new implementation of shadings, step 63.
DETAILS :

The new code is disabled with USE_LINEAR_COLOR_PROCS in gxshade4.h .

Fixing various problems with linear color :
1. Applyed compiler independent data types int32_t.
2. Implemented ends_from_fa with X-independent gradients,
   because the decomposiotion by color linearity breaks
   at coordinates which are not trapezoid corners.
   Added a comment to gxdevcli.h to document it.
3. X-gradient denominator was wrong.
4. X-gradient numerator left wrong from the step 58.
5. gx_default_fill_linear_color_scanline ignored the clipping by X.
6. A_fill_region missed the path bbox.
7. Implemented is_color_linear for a proper subdivision by color.
8. decompose_linear_color now applies a linear color
   instead fill_triangle_wedge. See a comment in fill_triangle_wedge.
   Also for mesh_padding.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4997 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevnfwd.c
@@ -770,7 +770,7 @@ gx_forward_include_color_space(gx_device *dev, gs_color_space *cspace,
 int 
 gx_forward_fill_linear_color_scanline(gx_device *dev, const gs_fill_attributes *fa,
 	int i, int j, int w,
-	const frac31 *c, const ulong *addx, const long *mulx, ulong divx)
+	const frac31 *c, const int32_t *addx, const int32_t *mulx, int32_t divx)
 {
     gx_device_forward * const fdev = (gx_device_forward *)dev;
     gx_device *tdev = fdev->target;
